LPS-44515 Extend new Base test case class in old test class and delete and override some abstract methods
