+++ portal-impl/test/integration/com/liferay/portal/util/BaseUserNotificationTestCase.java
@@ -19,8 +19,6 @@ import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.notifications.UserNotificationDefinition;
 import com.liferay.portal.kernel.test.JDKLoggerTestUtil;
 import com.liferay.portal.kernel.util.ArrayUtil;
-import com.liferay.portal.kernel.util.Constants;
-import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.model.BaseModel;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.model.User;
@@ -28,12 +26,9 @@ import com.liferay.portal.model.UserNotificationDelivery;
 import com.liferay.portal.model.UserNotificationDeliveryConstants;
 import com.liferay.portal.model.UserNotificationEvent;
 import com.liferay.portal.service.GroupLocalServiceUtil;
-import com.liferay.portal.service.ServiceContext;
-import com.liferay.portal.service.ServiceTestUtil;
 import com.liferay.portal.service.UserNotificationDeliveryLocalServiceUtil;
 import com.liferay.portal.service.UserNotificationEventLocalServiceUtil;
 import com.liferay.portlet.blogs.service.BlogsEntryLocalServiceUtil;
-import com.liferay.portlet.blogs.util.BlogsTestUtil;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -49,7 +44,7 @@ import org.junit.Test;
  * @author Roberto Díaz
  * @author Sergio González
  */
-public class BaseUserNotificationTestCase extends BaseMailTestCase {
+public abstract class BaseUserNotificationTestCase extends BaseMailTestCase {
 
 	@Before
 	@Override
@@ -313,18 +308,7 @@ public class BaseUserNotificationTestCase extends BaseMailTestCase {
 		Assert.assertEquals(0, userNotificationEventsJSONObjects.size());
 	}
 
-	protected BaseModel addBaseModel() throws Exception {
-		ServiceContext serviceContext = ServiceTestUtil.getServiceContext(
-			group.getGroupId());
-
-		serviceContext.setCommand(Constants.ADD);
-		serviceContext.setLayoutFullURL("http://localhost");
-		serviceContext.setScopeGroupId(group.getGroupId());
-
-		return BlogsTestUtil.addEntry(
-			user.getUserId(), ServiceTestUtil.randomString(), true,
-			serviceContext);
-	}
+	protected abstract BaseModel addBaseModel() throws Exception;
 
 	protected void deleteUserNotificationDeliveries() throws Exception {
 		UserNotificationDeliveryLocalServiceUtil.
@@ -410,21 +394,8 @@ public class BaseUserNotificationTestCase extends BaseMailTestCase {
 		return userNotificationEventJSONObjects;
 	}
 
-	protected void updateBaseModel(BaseModel baseModel) throws Exception {
-		ServiceContext serviceContext = ServiceTestUtil.getServiceContext();
-
-		serviceContext.setCommand(Constants.UPDATE);
-		serviceContext.setLayoutFullURL("http://localhost");
-		serviceContext.setScopeGroupId(group.getGroupId());
-
-		BlogsEntryLocalServiceUtil.updateEntry(
-			blogsEntry.getUserId(), blogsEntry.getEntryId(),
-			ServiceTestUtil.randomString(), blogsEntry.getDescription(),
-			blogsEntry.getContent(), 1, 1, 2012, 12, 00, true, true,
-			new String[0], blogsEntry.getSmallImage(),
-			blogsEntry.getSmallImageURL(), StringPool.BLANK, null,
-			serviceContext);
-	}
+	protected abstract void updateBaseModel(BaseModel baseModel)
+		throws Exception;
 
 	protected void updateUserNotificationDelivery(
 			int notificationType, int deliveryType, boolean deliver)
